OpenSea/requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==3.0.4
cloudscraper==1.2.58
discord-webhook==0.14.0
html5lib==1.1
idna==2.10
lxml==4.6.3
pyparsing==2.4.7
python-dotenv==0.19.0
requests==2.24.0
requests-toolbelt==0.9.1
six==1.16.0
soupsieve==2.2.1
urllib3==1.25.11
webencodings==0.5.1
